pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2epub"
version = "0.1.0"
description = "Convert Persian PDF text into RTL EPUB files."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "PDF2EPUB Author" }
]
dependencies = [
  "pymupdf>=1.22.0",
  "ebooklib>=0.18",
  "python-bidi>=0.4.2",
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
  "pdfplumber>=0.8.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
pdf2epub = "pdf2epub.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pdf2epub*"]
